fix: 修复钉钉同步 save-org-settings 2081 溢出，get-all-data 分页防截断

+++ CloudProgram/cloudfunctions/save-org-settings/OrgSettings.ts
@@ -16,7 +16,7 @@ export class OrgSettings {
     map.set('roleLabels', 'String');
     map.set('dingtalkClientId', 'String');
     map.set('dingtalkClientSecret', 'String');
-    map.set('dingtalkLastSyncAt', 'Integer');
+    map.set('dingtalkLastSyncAt', 'Long');
     map.set('dingtalkLastResult', 'String');
     return map;
   }
